changed all instances of double to a #define type of FLOAT ( = double) as set in Settings, so it can be easily changed later. Also a few minor re-names.

+++ Headers/Motion.h
@@ -14,7 +14,7 @@ namespace DualityEngine {
 
     struct Motion
     {
-        Motion(const double &velX, const double &velY, const double &velZ, const double &anvX, const double &anvY, const double &anvZ){
+        Motion(const FLOAT &velX, const FLOAT &velY, const FLOAT &velZ, const FLOAT &anvX, const FLOAT &anvY, const FLOAT &anvZ){
             velLinear = {velX, velY, velZ};
             velAngular = {anvX, anvY, anvZ};
         }
