+++ Headers/Spatial.h
@@ -14,7 +14,7 @@ namespace DualityEngine {
 
     struct Spatial
     {
-        Spatial(const double &posX, const double &posY, const double &posZ, const double &rotX, const double &rotY, const double &rotZ){
+        Spatial(const FLOAT &posX, const FLOAT &posY, const FLOAT &posZ, const FLOAT &rotX, const FLOAT &rotY, const FLOAT &rotZ){
             position = {posX, posY, posZ};
             rotation = {rotX, rotY, rotZ};
         }
